+++ CHANGELOG.md
@@ -12,9 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command line argument parsing
 - Configuration Options
 - Connection Opening
-- Connection Data Processing
-- Scenario File Parsing and Processing
-
+- Thread Controller
 
 ### Changed